trying fast build
stop building front end and reuse official one

+++ ci/build.sh
@@ -5,9 +5,8 @@ set -x
 usage() {
   base="$(basename "$0")"
   cat <<EOUSAGE
-usage: $base [-r] <arch>
-Install specific packages to build grafana for either armv6, armv7 or arm64
-Use -r for release package
+usage: $base <arch>
+Faster grafana build (armv6, armv7 or arm64) reusing offical assets
 Available arch:
   $base armv6
   $base armv7
@@ -15,17 +14,36 @@ Available arch:
 EOUSAGE
 }
 
+download_official_assets(){
+  if [ ! -f /tmp/grafana_${GRAFANA_VERSION}_amd64.deb ]; then
+    curl -L https://s3-us-west-2.amazonaws.com/grafana-releases/release/grafana_${GRAFANA_VERSION}_amd64.deb \
+         -o /tmp/grafana_${GRAFANA_VERSION}_amd64.deb
+    cd /tmp/ && dpkg-deb -R grafana_${GRAFANA_VERSION}_amd64.deb deb
+    mv /tmp/deb/DEBIAN /tmp/DEBIAN
+  fi
+  if [ ! -f /tmp/grafana-${GRAFANA_VERSION}.linux-x64.tar.gz ]; then
+    curl -L https://s3-us-west-2.amazonaws.com/grafana-releases/release/grafana-${GRAFANA_VERSION}.linux-x64.tar.gz \
+         -o /tmp/grafana-${GRAFANA_VERSION}.linux-x64.tar.gz
+    mkdir /tmp/tgz
+    cd /tmp/ && tar xf /tmp/grafana-${GRAFANA_VERSION}.linux-x64.tar.gz -C tgz
+  fi
+}
+
 install_phjs() {
   PHJSURL="https://github.com/fg2it/phantomjs-on-raspberry/releases/download/${PHJSV}"
   PHJS=/tmp/${ARM}/phantomjs
   mkdir -p /tmp/${ARM}
-  curl -sSL ${PHJSURL}/phantomjs -o ${PHJS}
-  chmod a+x ${PHJS}
+  if [ ! -f ${PHJS} ]; then
+    curl -sSL ${PHJSURL}/phantomjs -o ${PHJS}
+    chmod a+x ${PHJS}
+  fi
 }
 
 armv6_install_cross(){
-  cd /tmp
-  git clone https://github.com/fg2it/cross-rpi1b.git
+  if [ ! -d /tmp/cross-rpi1b/ ]; then
+    cd /tmp
+    git clone https://github.com/fg2it/cross-rpi1b.git
+  fi
   CROSSPATH="/tmp/cross-rpi1b/arm-rpi-4.9.3-linux-gnueabihf/bin/"
   CC=${CROSSPATH}/arm-linux-gnueabihf-gcc
 }
@@ -40,56 +58,87 @@ arm64_install_cross() {
   CC=aarch64-linux-gnu-gcc
 }
 
-build() {
+build_bin32() {
   cd $GOPATH/src/github.com/grafana/grafana
-  go run build.go                   \
-     -pkg-arch=${ARCH}              \
-     -goarch=${ARM}                 \
-     -cgo-enabled=1                 \
-     -cc=$CC                        \
-     -phjs=${PHJS}                  \
-     -includeBuildNumber=${includeBuildNumber} \
-         build                      \
-         pkg-deb
+  CGO_ENABLED=1 GOARCH=arm GOARM=$1 CC=$CC go build -ldflags "-w -X main.version=${GRAFANA_VERSION} -X main.commit=${commit} -X main.buildstamp=${buildstamp}" \
+           -o ./bin/grafana-server ./pkg/cmd/grafana-server
+  CGO_ENABLED=1 GOARCH=arm GOARM=$1 CC=$CC go build -ldflags "-w -X main.version=${GRAFANA_VERSION} -X main.commit=${commit} -X main.buildstamp=${buildstamp}" \
+           -o ./bin/grafana-cli ./pkg/cmd/grafana-cli
+}
+
+build_bin64() {
+  cd $GOPATH/src/github.com/grafana/grafana
+  CGO_ENABLED=1 GOARCH=${ARM} CC=$CC go build -ldflags "-w -X main.version=${GRAFANA_VERSION} -X main.commit=${commit} -X main.buildstamp=${buildstamp}" \
+           -o ./bin/grafana-server ./pkg/cmd/grafana-server
+  CGO_ENABLED=1 GOARCH=${ARM} CC=$CC go build -ldflags "-w -X main.version=${GRAFANA_VERSION} -X main.commit=${commit} -X main.buildstamp=${buildstamp}" \
+           -o ./bin/grafana-cli ./pkg/cmd/grafana-cli
+}
+
+fix_assets(){
+  cd $GOPATH/src/github.com/grafana/grafana
+  cp ./bin/grafana-server /tmp/deb/usr/sbin/
+  cp ./bin/grafana-server /tmp/tgz/grafana-${GRAFANA_VERSION}/bin/
+  cp ./bin/grafana-cli /tmp/deb/usr/sbin/
+  cp ./bin/grafana-cli /tmp/tgz/grafana-${GRAFANA_VERSION}/bin/
+  cp $PHJS /tmp/deb/usr/share/grafana/tools/phantomjs/
+  cp $PHJS /tmp/tgz/grafana-${GRAFANA_VERSION}/tools/phantomjs/
+}
+
+repack_assets(){
+  cd /tmp/tgz
+  mkdir -p /tmp/pkg/${ARM}
+  tar cfz /tmp/pkg/${ARM}/grafana-${GRAFANA_VERSION}.linux-${ARCH}.tar.gz grafana-${GRAFANA_VERSION}
+  cd /tmp/deb
+  fpm -t deb -s dir --description Grafana -C /tmp/deb                         \
+      --vendor Grafana --url https://grafana.com --license "Apache 2.0"       \
+      --maintainer contact@grafana.com                                        \
+      --config-files /etc/init.d/grafana-server                               \
+      --config-files /etc/default/grafana-server                              \
+      --config-files /usr/lib/systemd/system/grafana-server.service           \
+      --after-install /tmp/DEBIAN/postinst                                    \
+      --name grafana --version "${GRAFANA_VERSION}"                           \
+      --deb-no-default-config-files -a ${ARCH}                                \
+      --depends adduser --depends libfontconfig -p /tmp/pkg/${ARM} .
 }
 
 
-includeBuildNumber="true"
-if [ "$1" == "-r" ]; then
-  echo "Package for release"
-  includeBuildNumber="false"
-  shift
-fi
+GRAFANA_VERSION='5.0.3'
+cd $GOPATH/src/github.com/grafana/grafana
+commit=$(git rev-parse --short HEAD)
+buildstamp=$(git show -s --format=%ct)
 
-if (( $# != 1 )); then
-	usage >&2
-	exit 1
-fi
+#download_official_assets
 
-ARM="$1"
-
-case "$ARM" in
-  armv6)
-    PHJSV="v2.1.1-wheezy-jessie-armv6"
-    armv6_install_cross
-    ARCH="armhf"
-    ;;
-  armv7)
-    PHJSV="v2.1.1-wheezy-jessie"
-    armv7_install_cross
-    ARCH="armhf"
-    ;;
-  arm64)
-    PHJSV="v2.1.1-jessie-stretch-arm64"
-    arm64_install_cross
-    ARCH="arm64"
-    ;;
-  *)
-    echo >&2 'error: unknown arch:' "$ARM"
-    usage >&2
-    exit 1
-    ;;
-esac
-
-install_phjs
-build
+for ARM in "$@"
+do
+  case "$ARM" in
+    armv6)
+      PHJSV="v2.1.1-wheezy-jessie-armv6"
+      install_phjs
+      armv6_install_cross
+      ARCH="armhf"
+      build_bin32 6
+      ;;
+    armv7)
+      PHJSV="v2.1.1-wheezy-jessie"
+      install_phjs
+      armv7_install_cross
+      ARCH="armhf"
+      build_bin32 7
+      ;;
+    arm64)
+      PHJSV="v2.1.1-jessie-stretch-arm64"
+      install_phjs
+      arm64_install_cross
+      ARCH="arm64"
+      build_bin64
+      ;;
+    *)
+      echo >&2 'error: unknown arch:' "$ARM"
+      usage >&2
+      exit 1
+      ;;
+  esac
+  fix_assets
+  repack_assets
+done
